added a simple selection system for the combat buttons - currently just works for the two buttons but should be dynamic
this is the beginning of transfering selection power to the UI manager instead of combat manager - couple of functions are now public in combat manager

inventory ui can also be de/activated using space - this was surprisingly hard to do. Code could probably be simplified but adding a delay and making sure nothing could activate it from outside was a needed step for the UI feel.

Next is making overworld work