have foundation for the mercy system, where the player has a chance to flee or for the enemy to be spared if their health is to 75%.
currently there is a bug where the combatmode is only selected for one frame and immediately reverts back menu. Something is reverting it back, but unsure of where. May be the input messing with the selection of the mercy mode? All that is certain is that something is causing it to revert and it is not in the section handling mercy mode