act system working. fixed minor bugs and separated the combat sprites into separate things without adding them to the scene yet
added a section to add in dialogue for when the correct act is chosen. This should be expanded upon with different dialogue options in a list and fancy ui, but this can be added later.
Not too sure how to actually make different weaknesses as they are enum based, but they can likely be changed to strings with minimal work. Just a matter of what would be best for the rest of the project.

combat manager goes to the enemy behavior for the dialogue to throw up. The dialogue is wep/WEP and wfp/WFP which communicates with the weaknessExploitedPhrase and the WeaknessFailPhrase in the phase class. This retrieves the current phrase based on the phase number.

Phases can now be passed based on this act system or the combat system, based on the amount of health the enemy has as determined in EnemyBehavior.

Note: We now have to just adjust the actual values in the fight to a greater degree, although it might be tedious its kinda just part of what is expected for tuning the battles